Add support for loading script files from the filesystem in addition to the class path
